backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
pydantic>=2.7.0
pydantic-settings>=2.3.0
sqlalchemy[asyncio]>=2.0.30
asyncpg>=0.29.0
alembic>=1.13.0
redis[asyncio]>=5.0.0
python-jose[cryptography]>=3.3.0
bcrypt>=4.1.3
httpx>=0.27.0
razorpay>=1.4.1
python-dotenv>=1.0.0
pyyaml>=6.0.1
